Tian and I worked on networking
We have it sending and receiving information, but for some reason the
slave interface has started going blank...we tried to fix it but are a
bit stuck. Was thinking Peter could look at the delay if we finish this
soon. Can you spot anything wrong with the slave interface Brian?